requirements/base.txt
https://www.djangoproject.com/download/1.7c2/tarball/
argparse==1.2.1
django-debug-toolbar==1.2.1
psycopg2==2.5.3
pyjade==2.2.0
pyserial==2.7
pytz==2014.2
simplejson==3.3.3
six==1.6.1
sqlparse==0.1.11
wsgiref==0.1.2
